requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.8.0
exceptiongroup==1.2.0
feedparser==6.0.11
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.6
packaging==23.2
pgvecto-rs==0.1.4
playwright==1.40.0
priority==2.0.0
psycopg==3.1.16
psycopg-binary==3.1.16
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
Quart==0.19.4
requests==2.31.0
sgmllib3k==1.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
toml==0.10.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
